Moves some stuff around.
