+++ ui/revenuecatui/src/main/kotlin/com/revenuecat/purchases/ui/revenuecatui/components/modifier/Border.kt
@@ -18,9 +18,9 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.revenuecat.purchases.paywalls.components.properties.ColorInfo
 import com.revenuecat.purchases.paywalls.components.properties.ColorScheme
-import com.revenuecat.purchases.ui.revenuecatui.components.BorderStyle
-import com.revenuecat.purchases.ui.revenuecatui.components.ColorStyle
-import com.revenuecat.purchases.ui.revenuecatui.components.toColorStyle
+import com.revenuecat.purchases.ui.revenuecatui.components.property.BorderStyle
+import com.revenuecat.purchases.ui.revenuecatui.components.property.ColorStyle
+import com.revenuecat.purchases.ui.revenuecatui.components.property.toColorStyle
 
 @JvmSynthetic
 @Stable
